Add files via upload
This is the Python code that takes an input, such as the pressing of a keyboard space bar, and outputs an audio signal (playSignal) which can be converted to vibrations via the haptic engine. This code was modified and implemented in the final main code in imu_bluetooth.py so that instead of the audio being played when the space bar is pressed, audio is emitted upon sensing right hand movement.